// openmetadata-ui/src/main/resources/ui/src/generated/api/data/ontologyBulkSubmission.ts
export enum ExecutionMode {
    Background = "BACKGROUND",
    Synchronous = "SYNCHRONOUS",
}